+++ src/Response/Cli.php
@@ -3,7 +3,7 @@ namespace Zaek\Framy\Response;
 
 class Cli extends Response
 {
-    public function showError($errorCode)
+    public function showError($errorCode) : void
     {
         echo "Error: {$errorCode}\n";
     }
@@ -13,7 +13,7 @@ class Cli extends Response
         echo $this->output;
         /*
         if(!is_null($this->result)) {
-            echo "Application result:\n";
+            echo "App result:\n";
             echo "--------------------\n";
             if(is_string($this->result)) {
                 echo $this->result;
@@ -24,7 +24,7 @@ class Cli extends Response
             echo "\n";
         }
         if(!is_null($this->output)) {
-            echo "Application output:\n";
+            echo "App output:\n";
             echo "--------------------\n";
             if(is_string($this->output)) {
                 echo $this->output;